+++ drivers/frame_provider/decoder_v4l/vav1_t5d/vav1_t5d.c
@@ -10378,9 +10378,9 @@ static unsigned long run_ready(struct vdec_s *vdec, unsigned long mask)
 		(get_double_write_mode(hw) != 0x10)) {
 		int size = decoder_mmu_box_sc_check(ctx->bm.mmu, tvp);
 		hw->first_sc_checked = 1;
-		av1_print(hw, 0, "av1 cached=%d  need_size=%d speed= %d ms\n",
+		av1_print(hw, 0, "av1 cached=%d  need_size=%d speed= %lld ms\n",
 			size, (hw->need_cache_size >> PAGE_SHIFT),
-			(int)(get_jiffies_64() - hw->sc_start_time) * 1000/HZ);
+			(get_jiffies_64() - hw->sc_start_time) * (1000/HZ));
 #ifdef AOM_AV1_MMU_DW
 		/*!!!!!! To do ... */
 		if (get_double_write_mode(hw) & 0x20) {
